feat: reverse the order of create page flow
