+++ apps/studio/src/theme/components/index.ts
@@ -1,7 +1,5 @@
-import { Modal } from "./Modal"
 import { Table } from "./Table"
 
 export const components = {
-  Modal,
   Table,
 }
